Add deduplicate_column module
This module fixes the awful damages a pd.DataFrame.merge does if not
done properly; i.e. the situation in which duplicated column appears
with _x and _y suffixes (or a combination of them if several chained
merge went wrong)

This module allows to merge back values in the respective original
columns (i.e. those without nasty suffixes), dealing correctly the
cases in which multiple (or contraddicting) values values are found

Usage:
         from  YAstarMM.deduplicate_columns  import   (
             columns_to_keep, fix_duplicated_columns,
         )

( or from within the YAstarMM package )

         from          .deduplicate_columns  import   (
             columns_to_keep, fix_duplicated_columns,
         )