hmm: improve interleaving of logging records
Introduce a global lock over the logging which allows a process to
flush its messages; the other processes will collect a backlog of
their messages and wait for the lock in order to flush them.

Improved the update of the global counter in charlson_index module in
preparation for future improvements (e.g. like multiprocessing
parallelism)