hmm: add number of days spent in a state to observed variables
This allows to mitigate the strong imbalance between the values on the
main diagonal of the occurrences/transition matrix
(i.e. self-transitions or self-edges) and the other entries on the
same line.

Moreover consecutive records in a patient dataframe (i.e. duplicated
records) can now be dropped; these two improvements should reduce the
sparsity and the redundance of the data in the
training/validation/test sets.

Please pay attention that only non-static columns/features (e.g. not
like age) are considered when dropping duplicated records